details view impl

+++ src/main/java/app/repository/SequenceRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SequenceRepository extends PagingAndSortingRepository<Sequence,Long> {}
+public interface SequenceRepository extends PagingAndSortingRepository<Sequence,Integer> {}
